Move Image model to Blogging\Models namespace
