+++ packages/framework/src/Framework/Models/Pages/MarkdownPost.php
@@ -5,11 +5,11 @@ declare(strict_types=1);
 namespace Hyde\Framework\Models\Pages;
 
 use Hyde\Blogging\Models\Author;
+use Hyde\Blogging\Models\Image;
 use Hyde\Foundation\PageCollection;
 use Hyde\Framework\Concerns\BaseMarkdownPage;
 use Hyde\Framework\Contracts\FrontMatter\BlogPostSchema;
 use Hyde\Framework\Models\Support\DateString;
-use Hyde\Framework\Models\Support\Image;
 
 /**
  * Page class for Markdown posts.
