+++ packages/framework/tests/Feature/FindsContentLengthForImageObjectTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Hyde\Framework\Testing\Feature;
 
-use Hyde\Framework\Models\Support\Image;
+use Hyde\Blogging\Models\Image;
 use Hyde\Testing\TestCase;
 use Illuminate\Support\Facades\Http;
 
@@ -16,7 +16,7 @@ class FindsContentLengthForImageObjectTest extends TestCase
     /**
      * Unit test for the shorthand. Logic is tested in the rest of the case.
      *
-     * @covers \Hyde\Framework\Models\Support\Image::getContentLength
+     * @covers \Hyde\Blogging\Models\Image::getContentLength
      */
     public function test_image_helper_shorthand_returns_content_length()
     {
